+++ src/packages/property-editors/multi-url-picker/manifests.ts
